Mount the nth FAT filesystem
The key-value file autoboot.txt lists the location of the redundant FAT
filesystems holding the file config.txt.

The section [all] sets the FAT filesystem to read if the tryboot reboot
flag is unset; and the section [tryboot] sets the FAT filesystem to read
if the tryboot reboot flag is set.

The property boot_partition sets the index of the nth partition of type
FAT filesystem, staring by 1, and it does **NOT** indicate the raw index
in the partition table!

In short, boot_partition=1 is the FAT filesystem containing the file
autoboot.txt, boot_partition=2 and boot_partition=3 are typically the
first and the second FAT filesystems if no extra FAT filesystem
interleave between these three filesystems.

This runs lsblk(8)[1] to list for the FAT filesystems and to generate
the systemd mount unit configuration encoding information about the nth
FAT filesystem given by the bootloader value `boot_partition`.

[1]: https://linux.die.net/man/8/lsblk